requirements.txt
mlflow==1.2.0
catboost==0.26
boto3==1.17.90
python-dotenv==0.17.1
scikit-learn==0.22.1